refactor to abstract calc_grad over dtype and to use float32 in test

+++ test/testing_utils.py
@@ -1,7 +1,8 @@
 import torch
-
-def calc_grad(run_function, device='cpu'):
-    t = torch.randn(10, 10, requires_grad=True, device=device)
+    
+# run_function should already by on correct device
+def calc_grad(run_function, device='cpu', dtype=torch.float32):
+    t = torch.randn(10, 10, requires_grad=True, device=device, dtype=dtype)
 
     y = run_function(t).sum()
     y.backward()
